Adds invincibility to player when hit.
Other changes include splitting characters.py into player and goblin files, even
though goblin.py only has an Enemy class (for now). The game also now ends when
there are no more goblins to hunt.
